% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedFieldNames.R
\name{bedFieldNames}
\alias{bedFieldNames}
\title{bedFieldNames}
\usage{
bedFieldNames(n = 12)
}
\arguments{
\item{n}{Number of fields}
}
\description{
Provide standard field names of BED files
}
\details{
See \url{http://www.genome.ucsc.edu/FAQ/FAQformat.html#format1 for details}.
}
\examples{
bedFieldNames(6)

}

