% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrFreqs.R
\name{chrFreqs}
\alias{chrFreqs}
\title{chrFreqs}
\usage{
chrFreqs(gr, chrs)
}
\arguments{
\item{gr}{A GenomicRanges object.}

\item{chrs}{A character vector of chromosome names.}
}
\value{
A named character vector indicating the number of feature
counts per chromosome in a GenomicRanges object.  This is done by
running GRanges' seqnames() function, converting the output to a
table, and converting the table to a named character vector, according
to the names passed to the function.  This is useful to create a
matrix or a data frame from a list of genomic ranges.
}
\description{
Count features per chromosome
}
\examples{
if(require(rtracklayer, quietly = TRUE)) {
  gr <- system.file("extdata", "BED12_A.bed", package="smallCAGEqc") \%>\%
          rtracklayer::import.bed
  chrFreqs(gr, c("chr1", "chr2", "chrX", "chrY", "chrZ"))
}

}
