% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erccSpikeConcentration.R
\name{erccSpikeConcentration}
\alias{erccSpikeConcentration}
\title{erccSpikeConcentration}
\usage{

  erccSpikeConcentration(file = "http://tools.thermofisher.com/content/sfs/manuals/cms_095046.txt")
}
\arguments{
\item{file}{File name or URL where to find the \sQuote{cms_095046.txt} text file.
Defaults to the current URL on the Thermo Fisher website.}
}
\value{
A data frame representing the file \sQuote{cms_095046.txt} from the vendors
website.

The original column names are \sQuote{Re-sort ID}, \sQuote{ERCC ID}, \sQuote{subgroup},
\sQuote{concentration in Mix 1 (attomoles/ul)},
\sQuote{concentration in Mix 2 (attomoles/ul)}, \sQuote{expected fold-change ratio},
\sQuote{log2(Mix 1/Mix 2)}, but this not fit well for a R data frame.  Therefore,
they are renamed as: \sQuote{sortID}, \sQuote{erccID}, \sQuote{subgroup},
\sQuote{concMix1}, \sQuote{concMix2}, \sQuote{FC}, \sQuote{log2FC}.
}
\description{
Various data related to ERCC spikes
}
\details{
In quantitative transcriptome analysis, we often add synthetic RNA to the reaction for
quality control and normalisation, and the External RNA Controls Consortium (ERCC)
spikes are a popular choice, available commercially from Invitrogen (now Thermo Fisher).
In the commercial product, the spikes have different concentrations, covering six
orders of magnitude.  These concentrations are given in a text file on the vendor's
webiste, and I do not know if the file is freely redistributable, hence this function
to retreive the data from the Internet or a local file.
}
\seealso{
loadMoiraiStats
}

