% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldgmChamberWell.R
\name{fldgmChamberWell}
\alias{fldgmChamberWell}
\title{fldgmChamberWell}
\usage{
fldgmChamberWell(...)
}
\arguments{
\item{...}{Either chamber numbers (nummeric format) or 96-well
names (character format)}
}
\description{
Converts C1 array capture chamber numbers to 96-well names and vice-versa
}
\examples{
fldgmChamberWell("E04")
fldgmChamberWell(c("E04", "A01"))
fldgmChamberWell(3, 2, 1)
fldgmChamberWell(c(3,2,1))
fldgmChamberWell(fldgmChamberWell(c(3,2,1)))

}
