% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldgmConcentrationPlot.R
\name{fldgmConcentrationPlot}
\alias{fldgmConcentrationPlot}
\title{fldgmConcentrationPlot}
\usage{
fldgmConcentrationPlot(LIBS, scales = "fixed", group = ~Run)
}
\arguments{
\item{LIBS}{A metadata table, often called \sQuote{libs} table,
that contains at least a \sQuote{Concentration} and a \sQuote{Run}
column.  The table will be coerced as a data frame
(without checking names), so that Bioconductor's DataFrame
objects can also be used.#'}

\item{scales}{The \sQuote{scales} argument for the plot.
Default: \sQuote{fixed}.}

\item{group}{A formula to be passed to \sQuote{ggplot2}'s
\sQuote{facet_wrap} function.}
}
\description{
Plot histogram of cDNA concentrations after a C1 run
}
\examples{
# Example data
libs <- read.table(system.file("extdata/libs-with-all-metadata.tsv", package="smallCAGEqc"))

# Plot a single histogram
fldgmConcentrationPlot(libs)

# Faced by group
fldgmConcentrationPlot(libs, group = ~group)

# Test coercion to data frame.
if(require(S4Vectors))
  fldgmConcentrationPlot(DataFrame(libs))

}

