% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannot.R
\name{hannot}
\alias{hannot}
\title{hannot}
\usage{
hannot(annot, hierarchy = "default")
}
\arguments{
\item{annot}{A character vector of annotations.}

\item{hierarchy}{Either the character \dQuote{default}, or a data frame
       whith patterns in its first column and keywords in its second
       column.  The default data frame is picked when \dQuote{default} is
       passed as argument.

The default data frame is as follows (as of version 0.2.1).

\preformatted{
None               None
snoscan            other
antisense          antisense
intron,sense\\|    intron
exon,sense\\|      exon
lncRNA\\|          exon
miRNA\\|           miRNA
snRNA\\|           snRNA
snoRNA\\|          snoRNA
tRNA\\|            tRNA
}}
}
\value{
A vector of annotations as factors, where the levels are sorted in
the hierarchical order.  For instance, with the data frame above, if an
annotation matches for the keywords \sQuote{miRNA} and \sQuote{snoRNA},
then the reported one will be \sQuote{snoRNA}.
}
\description{
Hierarchical annotation of the CAGE clusters
}
\details{
Parses a vector of annotations in character format an returns a vector of
annotations as factors, where one single keyword has been selected from a
provided list or the default list (see below).
}
\examples{
x <- c('exon,sense|foo', 'foo', 'antisense of bar', 'intron,sense|bar'
      , 'antisense of baz;snoRNA|SNORA70')
hannot(x)
cbind(x, as.character(hannot(x)))

}
\seealso{
Other smallCAGEqc annotation functions: \code{\link{hierarchAnnot}},
  \code{\link{mapStats}}, \code{\link{plotAnnot}}
}

