% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchAnnot.R
\name{hierarchAnnot}
\alias{hierarchAnnot}
\title{hierarchAnnot}
\usage{
hierarchAnnot(annot)
}
\value{
The annotation table with the hierarchical annotation in a column
called \code{class}.
}
\description{
Hierarchical annotation of the CAGE clusters
}
\details{
Parse an annotation data frame and retains only one annotation,
in the order \sQuote{promoter}, \sQuote{exon}, \sQuote{intron}
and \sQuote{unknown}.
}
\seealso{
\code{\link{loadLogs}}

Other smallCAGEqc annotation functions: \code{\link{hannot}},
  \code{\link{mapStats}}, \code{\link{plotAnnot}}
}

