% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imageRenameCellomicsC1.R
\name{imageRenameCellomicsC1}
\alias{imageRenameCellomicsC1}
\title{imageRenameCellomicsC1}
\usage{
imageRenameCellomicsC1(C1ID, directory, fileExtension = "C01")
}
\arguments{
\item{C1ID}{Run ID, for example "1772-066-262".}

\item{directory}{Directory containing raw files.}

\item{fileExtension}{File extension for the raw files.  Defaults to "C01".}
}
\description{
Rename raw image files into 96-well format.
}
\details{
Due to filesystem case-insensitivity, it is not possible to have 96
different identifiers in one Cellomics run.  Therefore, two runs
are made with arbitrary names, and the files are renamed with this
script to reflect coordinates in 96-well format.

WARNING: This function renames files and deletes directory; use it at
}
\author{
Michael Böttcher and Charles Plessy
}

