% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMoiraiStats.R
\name{loadMoiraiStats}
\alias{loadMoiraiStats}
\title{loadMoiraiStats}
\usage{
loadMoiraiStats(multiplex, summary, pipeline, ercc = FALSE)
}
\arguments{
\item{multiplex}{Optional. Path to a \sQuote{multiplex} file.}

\item{summary}{Optional. Path to a \sQuote{summary} file.}

\item{pipeline}{Optional. Version string identifying the pipeline used to process the data.}

\item{ercc}{Optional. [Experimental] Return ERCC spike counts.}
}
\value{
Returns a data frame with one row per sample, and the following columns (if the
corresponding data is available).
\enumerate{
 \item{samplename} {Sample identifier (factor)}
 \item{extracted} {Number of extracted reads}
 \item{tagdust} {Number of reads containing oligonucleotide artefacts}
 \item{spikes} {Number of reads overlaping with the reference spike sequences}
 \item{rdna} {Number of reads overlaping with the reference ribosomal DNA sequences}
 \item{mapped} {Number of reads aligned to the reference genome}
 }
 
 Alternatively, returns ERCC spike counts when \sQuote{ercc} is set to \sQuote{TRUE}.
}
\description{
Load mapping statistics from log files
}
\details{
Loads mapping counts and other statistics produced during processing. This
function deprecates \code{loadLogs("moirai")}.  It loads data from a summary file
and a multiplex file.  When their path is not given by \code{multiplex} and
\code{summary}, they will be searched at fixed locations in the
\code{PROCESSED_DATA} directory using the \code{LIBRARY} variable.

\code{loadMoiraiStats} will recognise the \sQuote{nano-fluidigm}
or the \sQuote{nanoCAGE2} Moirai users, or fail.  For the \sQuote{nano-fluidigm}
user, the samples are sorted by numbers and associated to sorted well names, from
A01, A02, ..., to H11 and H12.
}
\examples{
loadMoiraiStats(
    summary   = system.file("extdata/summary.txt", package="smallCAGEqc")
  , multiplex = system.file("extdata/samplename_to_sampleid.txt", package="smallCAGEqc")
  , pipeline  = "OP-WORKFLOW-CAGEscan-short-reads-v2.0" )
        
libs$group <- libs$samplename \%>\% sub("Run._", "", .) \%>\% substr(1,1) \%>\% factor
}
\seealso{
\code{\link{hierarchAnnot}}, code{\link{loadLogs}}, \code{\link{mapStats}}
}

