% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoCAGE2017barcodes.R
\docType{data}
\name{nanoCAGE2017barcodes}
\alias{nanoCAGE2017barcodes}
\title{96 hexameric barcode sequences for sample indexing}
\format{A data frame with one column named \code{barcodes}.}
\usage{
nanoCAGE2017barcodes
}
\description{
Sequence of the 96 barcodes used in Poulain \emph{et al.}, 2017. NanoCAGE
libraries can index samples by inserting barcodes in the template-switching
oligonucleotide.  A list of 96 barcodes, was published in Poulain \emph{et
al.}, 2017.  The barcodes were selected to avoid palindromes, polynucleotide
tracts, and sequence similarity.
}
\references{
Poulain S, Kato S, Arnaud O, Morlighem JÉ, Suzuki M, Plessy C,
  Harbers M. NanoCAGE: A Method for the Analysis of Coding and Noncoding
  5'-Capped Transcriptomes. Methods Mol Biol. 2017 1543:57-109
}
\author{
Jean-Étienne Morlighem, Charles Plessy
}
\keyword{datasets}
