% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaGeneLibs.R
\name{pcaCompOrientation}
\alias{pcaCompOrientation}
\title{pcaCompOrientation}
\usage{
pcaCompOrientation(compPca)
}
\arguments{
\item{compPca}{The vector of variable scores for a given principal component.}
}
\value{
A boolean object is returned.
}
\description{
Returns \code{TRUE} if the principal component outputed from \code{dudi.pca} 
object (package \sQuote{ade4}) is mostly oriented towards positive values ; 
else returns \code{FALSE}.
}
\details{
Take the principal component vector and check which extremity value is the farthest from 0.
}
\examples{
library(ade4)
data("doubs")
pcaCompOrientation(dudi.pca(doubs$env, scannf = F, nf = 3)$co[, 1])
pcaCompOrientation(dudi.pca(doubs$env, scannf = F, nf = 3)$co[, 3])

}
\seealso{
\code{\link{ade4::dudi.pca}}
}

