% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaGeneLibs.R
\name{pcaPlots}
\alias{pcaPlots}
\title{pcaPlots}
\usage{
pcaPlots(pcaAde4, cellcol = NULL, size = 1, transparency = 0.6,
  optim = T)
}
\arguments{
\item{pcaAde4}{An object outputed from a \code{dudi.pca} object (package \sQuote{ade4}).}

\item{cellcol}{A vector of colors regarding the individuals (cells). If not provided, all points will be black.}

\item{size}{An integer designating the size of the text on the plots.}

\item{transparency}{A numeric (0-1) designating the transparency of the text to display on plots.}
}
\value{
A list of plots of \sQuote{genes} (variables in PCA) or \sQuote{cells} 
(individuals or samples). Each element contains a list of plots of a variety of combination of axis.
}
\description{
Take the object outputed from a \code{dudi.pca} object (package \sQuote{ade4}) and
returns a list of plots of the genes and cells processed by the PCA according to a
variety of axis combination.
}
\details{
Take the object outputed from a \code{dudi.pca} object (package \sQuote{ade4}) and
create a plot for genes (variables) and cells (individuals) according to differents
combination of axis. If axis 3,2 plot has its absciss coordinates reversed to fit
a more visual configuration when plotting all 3 dimensions on three side-by-side graphs. 
For now can only process PCA with 3 axis (\code{nf=3} as argument of \code{dudi.pca}).
}
\examples{
library(ade4)
library(adegraphics)
data("doubs")
pcaPlots(dudi.pca(doubs$env, scannf = F, nf = 3))$genes$ax12
pcaPlots(dudi.pca(doubs$env, scannf = F, nf = 3), cellcol = rainbow(nrow(doubs$env)))$cells$ax12

}
\seealso{
\code{\link{ade4::dudi.pca}}
}

