% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate96.R
\name{plate96}
\alias{plate96}
\title{plate96}
\usage{
plate96(from, to, by = "row")
}
\arguments{
\item{from}{Well coordinates.}

\item{to}{Well coordinates.}

\item{by}{By \code{row} (default) or \code{column}.}
}
\description{
Helper tool for 96-well plate coordinates
}
\note{
\code{by="column"} is not yet implemented.
}
\examples{

plate96("A01", "B08")
plate96("C01", "B08")

}

