% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnnot.R
\name{plotAnnot}
\alias{plotAnnot}
\title{Plot annotation statistics}
\usage{
plotAnnot(LIBS, SCOPE, TITLE, GROUP = "default", customScope = NULL,
  normalise = TRUE)
}
\arguments{
\item{LIBS}{A table with columns named \code{promoter}, \code{exon},
\code{intron}, \code{mapped}, \code{extracted}, \code{rdna}, and
\code{tagdust}, that will be passed to the \code{mapStats} function.
The table will be coerced to a data frame, so that Bioconductor's
DataFrame objects can be used.}

\item{SCOPE}{The value on which to normalise (see the plotAnnot vignette).}

\item{TITLE}{The title of the plot.}

\item{customScope}{A function passed to \code{\link{mapStats}} for the
definition of custom scopes}

\item{normalise}{Whether to normalise or not. Default: TRUE.}
}
\description{
Using a data frame containing mapping statistics in counts, plot the data as
percentages in stacked barplots.
}
\details{
Stacked barplots with error bars inspired from
<http://stackoverflow.com/questions/10417003/stacked-barplot-with-errorbars-using-ggplot2>.
See <http://www.biomedcentral.com/1471-2164/14/665/figure/F1> for example.
}
\examples{
example(loadMoiraiStats)
p <- plotAnnot(libs, 'qc', 'Here is the title')
print(p)
p + ggplot2::theme_bw()
ggplot2::theme_set(ggplot2::theme_bw()) ; p
plotAnnot(libs, 'qc', 'Same, non-normalised', normalise = FALSE)
# Test coercion to data frame
if(require("S4Vectors"))
  plotAnnot(DataFrame(libs), 'qc', 'Here is the title')
}
\seealso{
\code{\link{loadLogs}}

Other smallCAGEqc annotation functions: \code{\link{hannot}},
  \code{\link{hierarchAnnot}}, \code{\link{mapStats}},
  \code{\link{ranges2annot}}, \code{\link{ranges2genes}}
}
