% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaGeneLibs.R
\name{plotHTB}
\alias{plotHTB}
\title{plotHTB}
\usage{
plotHTB(orderedCompPca, comp, nbDispGenes = 25)
}
\arguments{
\item{orderedCompPca}{A \code{data.frame} outputed from \sQuote{pcaCompGenesList}.}

\item{comp}{An integer designating the principal component we want to plot the barplot on.}

\item{nbDispGenes}{An integer designating how many genes we want to display.}
}
\value{
Display a barplot.
}
\description{
Return a barplot of the first and last genes revealed by the principal 
component (from \code{dudi.pca}) object (package \sQuote{ade4}).
}
\details{
Take the \code{data.frame} outputed from \sQuote{pcaCompGenesList} and display on 
the same barplot, the \code{head(., nbDispGenes)} and \code{tail(., nbDispGenes)} with 
the gene names below each bar.
}
\examples{
plotHTB(pcaCompGenesList(dudi.pca(dataFrame)$co, 1), 1, 10)
  
}
\seealso{
\code{\link{ade4::dudi.pca}} \code{\link{pcaCompGenesList}}
}

