% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanabiPlot.R
\name{points.hanabi}
\alias{points.hanabi}
\title{Add a final point in hanabi plots.}
\usage{
\method{points}{hanabi}(h, ...)
}
\arguments{
\item{h}{The hanabi object.}

\item{...}{Other parameters passed to the generic points function}
}
\description{
Will only add a point for the final, non-subsampled value of each
sample of in a hanabi object.
}
\seealso{
hanabi, plot.hanabi
}

