% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges2.R
\name{ranges2annot}
\alias{ranges2annot}
\title{ranges2annot}
\usage{
ranges2annot(ranges, annot, showClasses = NULL)
}
\arguments{
\item{ranges}{The Genomics Ranges object, for example extracted from a
RangedSummarizedExperiment object with the \code{rowRanges}
command.}

\item{annot}{A Genomics Ranges object providing enough information for
inferring promoters, exons and introns.  Typically GENCODE.}
}
\value{
A factor of same length as the GRanges object, indicating if the
        interval is promoter, exon, intron or unknown.
}
\description{
hierarchical annotation of a Genomic Range as promoter, exon or intron.
}
\examples{
\dontrun{
rowRanges(l1)$annotation <- ranges2annot(rowRanges(l1), gff)
}

}
\seealso{
Other smallCAGEqc annotation functions: \code{\link{hannot}},
  \code{\link{hierarchAnnot}}, \code{\link{mapStats}},
  \code{\link{plotAnnot}}, \code{\link{ranges2genes}}
}

