% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges2.R
\name{ranges2genes}
\alias{ranges2genes}
\title{ranges2genes}
\usage{
ranges2genes(ranges, genes)
}
\arguments{
\item{ranges}{The Genomics Ranges object, for example extracted from a
RangedSummarizedExperiment object with the \code{rowRanges}
command.}

\item{annot}{A Genomics Ranges object providing enough information for
finding gene symbols.  Typically GENCODE.}
}
\value{
A character vector of same length as the GRanges object, indicating
        one gene symbol or a comma-separated list of gene symbols for each
        range.
}
\description{
Assign gene symbol(s) to Genomic Ranges.
}
\examples{
\dontrun{
rowRanges(l1)$genes  <- ranges2genes(rowRanges(l1), gff)
}

}
\seealso{
Other smallCAGEqc annotation functions: \code{\link{hannot}},
  \code{\link{hierarchAnnot}}, \code{\link{mapStats}},
  \code{\link{plotAnnot}}, \code{\link{ranges2annot}}
}
