% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlock.R
\name{setBlock}
\alias{setBlock}
\title{setBlock}
\usage{
setBlock(plate, block, what, value)
}
\arguments{
\item{plate}{A data frame representing a multiwell plate ("long format").}

\item{block}{Coordinates of a rectangular block, or a vector of coordinates.}

\item{what}{A column name in the plate data frame.}

\item{value}{The value to insert or update with.}
}
\description{
Define properties for a rectangular block in a 96- or 384-well plate.
}
\details{
Update a data frame representing a multiwell plate, by setting a given value
for all wells in a block or a list of blocks defined by the well coordinates
of their upper-left and bottom-right corners.
}
\examples{
\dontrun{
plate <- setBlock(plate, "A01~C04", "dNTP", 0.5)

# Same with magrittr
plate \%<>\% setBlock("A01~C04", "dNTP", 0.5)

# Chained updates with magrittr
plate \%<>\%
  setBlock("A01~C04", "dNTP", 0.5) \%>\%
  setBlock("A01~C04", "Mg",   3.0)
}

p <- data.frame(Well = platetools::num_to_well(1:96))
p$Row <- sub("..$", "", p$Well) \%>\% factor
p$Col <- sub("^.",  "", p$Well) \%>\% as.numeric \%>\% factor

p \%>\% setBlock("A01~B02", "dNTP", 0.5) \%>\% head
p \%>\% setBlock(c("A01~B02", "D01~D02"), "dNTP", 0.5) \%>\% head


}

