% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallCAGEqc.R
\docType{package}
\name{smallCAGEqc}
\alias{smallCAGEqc}
\alias{smallCAGEqc-package}
\title{smallCAGEqc}
\description{
A package to quickly generate reports about CAGE libraries.
}
\details{
Ad-hoc functions to QC shallow-sequenced libraries, for instance from MiSeq.

Under construction
}
