% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInEnv.R
\name{exportInEnv}
\alias{exportInEnv}
\title{exportInEnv}
\usage{
exportInEnv(...)
}
\arguments{
\item{...}{Names of variables.}
}
\value{
Returns shell commands that, would have the same effect.  This is useful when
cut-pasting from knitR workflows that use R to change the environment and then
run some shell chunks.
}
\description{
Takes R variables and exprort them as environment variables
}
\details{
This function was developed for analysis pipelines running R and shell chunks
with knitR.  Changing the environment in one shell chunk has no effect on the
next ones, but changing it in a R chunk has a global effect.

This function uses \code{.Internal(Sys.setenv()} because its syntax was more
friendly for automation.
}
\examples{
foo <- "bar"
toto <- "tata"
exportInEnv("foo", "toto")
}

