% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldgmConcentrationPlot.R
\name{fldgmConcentrationPlot}
\alias{fldgmConcentrationPlot}
\title{fldgmConcentrationPlot}
\usage{
fldgmConcentrationPlot(LIBS, scales = "fixed", group = ~Run)
}
\arguments{
\item{LIBS}{A \sQuote{libs} table, that contains at least a
\sQuote{Concentration} and a \sQuote{Run} column}

\item{scales}{The \sQuote{scales} argument for the plot.
Default: \sQuote{fixed}.}

\item{group}{A formula to be passed to \sQuote{ggplot2}'s
       \sQuote{facet_wrap} function.}
}
\description{
Plot histogram of cDNA concentrations after a C1 run
}

