% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanabiPlot.R
\name{hanabi}
\alias{hanabi}
\title{hanabi}
\usage{
hanabi(expr_data, npoints = 20, step = 0.75, from = NULL)
}
\arguments{
\item{expr_data}{An expression table where columns are samples and rows
are features such as genes, TSS, etc, or a vector of counts.}

\item{npoints}{The maximum number of rarefactions per sample.}

\item{step}{Subsample sizes are calculated by taking the largest sample
and multiplying it by the step "npoints" times.}

\item{from}{Add one sample size (typically "0") in order to extend the
plot on the left-hand side.}
}
\value{
A list-based object of class "hanabi".
}
\description{
Rarefy data at multiple sample sizes, in preparation for plotting.
}
\details{
The computation can be long, so the steps of rarefaction and plotting
are kept separate.

The input must be a data frame, or a vector or a matrix, which
         will be coerced into a matrix.  The data must be counts (tag
         counts, molecule counts, ...).
}
\examples{

bedFiles <- system.file(package = "smallCAGEqc", "extdata") \%>\%
              list.files("*BED", full.names = TRUE)
bed <- loadBED12(bedFiles)
rar <- tapply(bed$score, bed$library, hanabi, from = 0) \%>\%
         structure(class = "hanabi")  # tapply discards the class !
hanabiPlot(rar, GROUP = levels(bed$library))
}
\seealso{
`[.hanabi`, as.list.hanabi, and vegan::rarecurve.

Other Hanabi functions: \code{\link{hanabiPlot}},
  \code{\link{plot.hanabi}}, \code{\link{points.hanabi}}
}
