% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanabiPlot.R
\name{hanabiPlot}
\alias{hanabiPlot}
\title{hanabiPlot}
\usage{
hanabiPlot(RAR, S, GROUP = NULL, legend.pos = "topleft", pch = 1,
  col = "black", ...)
}
\arguments{
\item{RAR}{A rarefaction table, or a hanabi object.}

\item{S}{A vector of subsample sizes.}

\item{GROUP}{A vector grouping the samples.  Coerced to factor.}

\item{legend.pos}{Position of the legend, passed as "x" parameter to the
"legend" function.}

\item{pch}{Plot character at the tip of the lines.}

\item{col}{A vector of colors}

\item{...}{Further arguments to be passed to the first plot function,
that plots the empty frame.}
}
\description{
Plot feature discovery curves
}
\details{
Plots the number of features (genes, transcripts, ...) detected for a
given number of counts (reads, unique molecules, ...).  Each library is
sub-sampled by rarefaction at various sample sizes, picked to provide
enough points so that the curves look smooth.  The final point is plotted
as an open circle, hence the name "hanabi", which means fireworks in
Japanese.

The rarefactions take time to do, so this step is done by a separate
function, so that the result is easily cached.
}
\examples{
\dontrun{
hanabi(genes, npoints = 20, step = 0.8, from = 0) \%>\% hanabiPlot
hanabi(genes, npoints = 20, step = 0.9) \%>\% hanabiPlot
}
bedFiles <- system.file(package = "smallCAGEqc", "extdata") \%>\%
              list.files("*BED", full.names = TRUE)
bed <- loadBED12(bedFiles)
rar <- tapply(bed$score, bed$library, hanabi, from = 0) \%>\%
         structure(class = "hanabi")  # tapply discards the class !
hanabiPlot(rar, GROUP = levels(bed$library))
hanabiPlot(rar, GROUP = levels(bed$library), col=c("red", "green", "blue"))
hanabiPlot(rar, col="purple")

}
\seealso{
vegan, plot.hanabi, hanabi

Other Hanabi functions: \code{\link{hanabi}},
  \code{\link{plot.hanabi}}, \code{\link{points.hanabi}}
}

