% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/llPostProcess.R
\name{llPostProcess}
\alias{llPostProcess}
\title{llPostProcess}
\usage{
llPostProcess(TABLE, TYPE = "")
}
\arguments{
\item{TABLE}{A table output by loadLogs}

\item{TYPE}{A type of post-processing}
}
\description{
Post-process the output of loadLogs.
}
\details{
If the TYPE is "nano-fluidigm", the generic "RunA" and "RunB" names are
replaced with the C1 run IDs found in the "RunA" and "RunB" variables,
which must exist.  These new sample names are then used are row names.
They are parsed to produce new columns for "Well", "row" and "column"
IDs in the 96-well plate format.
}
\seealso{
loadLogs
}

