% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapStats.R
\name{mapStats}
\alias{mapStats}
\title{mapStats}
\usage{
mapStats(libs, scope = c("all", "annotation", "counts", "mapped", "qc"),
  group = "default")
}
\arguments{
\item{libs}{A data frame with columns named \code{promoter}, \code{exon}, \code{intron}
\code{mapped}, \code{extracted}, \code{rdna}, and \code{tagdust}.}

\item{scope}{The value on which to normalise. \dQuote{all} and \dQuote{qc}
       normalise on the number of extracted tags, \dQuote{annotation} on the
       number of aligned tags, \dQuote{mapped} on the number of aligned tags
       and \dQuote{counts} on the transcript counts.}
}
\value{
Returns mean and standard deviation of normalised mapping statistics, plus absolute
positions for the error bars.
}
\description{
Process mapping statistics
}
\details{
Using a data frame containing mapping statistics in counts, transform the data in
percentages that can be used for stacked barplots.

See http://stackoverflow.com/questions/10417003/stacked-barplot-with-errorbars-using-ggplot2 about stacked barplot.

The \dQuote{mapped} and \dQuote{counts} scopes assume that transcript counts are available.
}
\seealso{
\code{\link{hierarchAnnot}}, \code{\link{loadLogs}}, \code{\link{plotAnnot}}
}

