\name{mapStats}
\alias{mapStats}
\title{Process mapping statistics}
\description{
Using a data frame containing mapping statistics in counts, transform the data in
percentages that can be used for stacked barplots.
}
\usage{
mapStats(libs)
}
\arguments{
  \item{libs}{
A data frame with columns named \code{promoter}, \code{exon}, \code{intron}
\code{Mapped}, \code{Extracted}, \code{rRNA}, and \code{TagDust}.
}
}
\details{
To Do
}

\value{
Returns data in percentages that can be used for stacked barplots.
}

\author{
Charles Plessy <plessy@riken.jp>
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}

## The function is currently defined as
\keyword{ ~CAGE }
\keyword{ ~annotation }% __ONLY ONE__ keyword per line
