% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapStats.R
\name{mapStats}
\alias{mapStats}
\title{mapStats}
\usage{
mapStats(libs, scope = c("all", "annotation", "counts", "mapped", "qc"),
  group = "default")
}
\arguments{
\item{libs}{A data frame with columns named \code{promoter}, \code{exon}, \code{intron}
\code{mapped}, \code{extracted}, \code{rdna}, and \code{tagdust}.}

\item{scope}{The value on which to normalise. \dQuote{all} and \dQuote{qc}
normalise on the number of extracted tags, \dQuote{annotation} on the
number of aligned tags, \dQuote{mapped} on the number of aligned tags
and \dQuote{counts} on the transcript counts.}

\item{group}{A vector of factors defining groups in the data.  By default,
the \dQuote{group} column of the \dQuote{libs} table.}
}
\value{
Returns a data frame with mean and standard deviation of normalised mapping statistics,
plus absolute positions for the error bars.  The first column, \code{group}, is
a vector of factors sorted with the \code{gtools::mixedorder} function.
}
\description{
Process mapping statistics
}
\details{
Using a data frame containing mapping statistics in counts, transform the data in
percentages that can be used for stacked barplots.

See http://stackoverflow.com/questions/10417003/stacked-barplot-with-errorbars-using-ggplot2 about stacked barplot.

The \dQuote{mapped} and \dQuote{counts} scopes assume that transcript counts are available.
}
\examples{
libs <- read.csv( system.file("extdata", "libs.csv", package = "smallCAGEqc")
                , row.names = 1)
mapStats(libs, "qc", libs$Error)

}
\seealso{
\code{\link{loadLogs}}

Other smallCAGEqc.annotation.functions: \code{\link{hannot}},
  \code{\link{hierarchAnnot}}, \code{\link{plotAnnot}}
}

