% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaGeneLibs.R
\name{pcaCompGenesList}
\alias{pcaCompGenesList}
\title{pcaCompGenesList}
\usage{
pcaCompGenesList(pcaAde4co, comp)
}
\arguments{
\item{pcaAde4co}{The \code{data.frame} of variable scores output from the \sQuote{co} 
element of a \code{dudi.pca} object (package \sQuote{ade4}).}

\item{comp}{An integer designating the principal component we want to sort .
the \code{data.frame} on.}
}
\value{
Only if 3 principal components were kept, returns a sorted \code{data.frame} based 
on principal component number \code{comp}. First column is the gene name. Second to 
fourth columns are score values of the gene on each principal components.
}
\description{
Returns \code{TRUE} if the principal component output from \code{dudi.pca} 
object (package \sQuote{ade4}) is mostly oriented towards positive values ; 
else returns \code{FALSE}.
}
\details{
Check if there are 3 principal components. If TRUE, then create a \sQuote{geneNames} column
based on the rownames of the \sQuote{pcaAde4co} \code{data.frame}. Then rearrange the column 
for \sQuote{geneNames} to be at the beginning. Then order the current \code{data.frame} based 
on column \code{comp}.
}
\examples{
pcaCompGenesList(dudi.pca(dataFrame)$co, 1)
  
}
\seealso{
\code{\link{ade4::dudi.pca}}
}

