% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotAnnot.R
\name{plotAnnot}
\alias{plotAnnot}
\title{Plot annotation statistics}
\usage{
plotAnnot(LIBS, SCOPE, TITLE, GROUP = "default")
}
\arguments{
\item{LIBS}{A data frame with columns named \code{promoter}, \code{exon}, \code{intron}
\code{mapped}, \code{extracted}, \code{rdna}, and \code{tagdust}, that will be
passed to the \code{mapStats} function.}

\item{SCOPE}{The value on which to normalise (see \code{mapStats}).}

\item{TITLE}{The title of the plot.}
}
\description{
Using a data frame containing mapping statistics in counts, plot the data as
percentages in stacked barplots.
}
\details{
Stacked barplots with error bars inspired from
<http://stackoverflow.com/questions/10417003/stacked-barplot-with-errorbars-using-ggplot2>.
See <http://www.biomedcentral.com/1471-2164/14/665/figure/F1> for example.
}
\examples{
libs <- read.csv( system.file("extdata", "libs.csv", package = "smallCAGEqc")
                , row.names = 1)
p <- plotAnnot(libs, 'qc', 'Here is the title', libs$Error)
\dontrun{
print(p)
p + ggplot2::theme_bw()
ggplot2::theme_set(ggplot2::theme_bw()) ; p
}
}
\seealso{
\code{\link{hierarchAnnot}}, \code{\link{loadLogs}}, \code{\link{mapStats}}
}

