% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupForMoirai.R
\name{setupForMoirai}
\alias{setupForMoirai}
\title{setupForMoirai}
\usage{
setupForMoirai(LIBRARY, MOIRAI_USER, MOIRAI_PROJECT, MOIRAI_STAMP = "latest",
  GENOME = c("hg19", "hg38"))
}
\arguments{
\item{LIBRARY}{A name identifying a library given in Moirai.}

\item{MOIRAI_USER}{A Moirai user name (part of the file path).}

\item{MOIRAI_PROJECT}{A Moirai project name (part of the file path).
Note that some Moirai project names look like user names and
vice versa.}

\item{MOIRAI_STAMP}{A Moirai time stamp.}

\item{GENOME}{A genome identifier, to get default annoation and gene definition.}
}
\value{
Prints shell commands that, would have the same effect.  This is useful when
cut-pasting from knitR workflows that use R to change the environment and then
run some shell chunks.

Returns PROCESSED_DATA.
}
\description{
Exports environment variables to find Moirai files.
}
\details{
LOCAL FUNCTION WITH A LOT OF HARDOCODING.  If no timestamp is given,
it will look for the latest processing.

Using LIBRARY, MOIRAI_USER and MOIRAI_PROJECT, \code{setupForMoirai} will find
the directory containing the last Moirai run for that library, and export its
path in the environment variable PROCESSED_DATA.

It will also set the environment variables GROUP_SHARED, GENE_SYMBOLS and
ANNOTATION, to default values.
}
\examples{

## PROCESSED_DATA <- setupForMoirai(
##   LIBRARY        = '140609_M00528_0024_000000000-A8CBW',
##   MOIRAI_USER    = 'nano-fluidigm',
##   MOIRAI_PROJECT = 'C1_CAGE' )
}
\seealso{
exportInEnv
}

