\name{smallCAGEqc-package}
\alias{smallCAGEqc-package}
\alias{smallCAGEqc}
\docType{package}
\title{
Functions to QC shallow-sequenced libraries.
}
\description{
Ad-hoc functions to QC shallow-sequenced libraries, for instance from MiSeq.
}
\details{
Under construction
}
\author{
Charles Plessy <plessy@riken.jp>
}
\references{
To do
}
\keyword{ package }
\keyword{ CAGE }
\seealso{
}
\examples{
}
