% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TPM.R
\name{TPM}
\alias{TPM}
\title{TPM}
\usage{
TPM(clusters)
}
\arguments{
\item{clusters}{A data frame of tag counts, where each row is a cluster and each
       column is a library.}
}
\value{
Returns a data frame expression values normalised in tags per million,
where each row is a cluster and each column is a library.
}
\description{
Tags Per Million
}
\details{
Takes a data frame of tag counts and return a data frame of tags per million.

Ensures that the column names of the returned data frame are identical to the input.
}
\examples{
TPM(data.frame(A=c(1,0,3), b=c(2,5,9)))

# Should warn about NaN
# TPM(data.frame(A=c(1,0,3), b=c(2,5,9), c=c(0,0,0)))
}

