% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennAB.R
\name{VennAB}
\alias{VennAB}
\title{VennAB}
\usage{
VennAB(A_NAME, A, B_NAME, B, MAIN, SUB)
}
\arguments{
\item{A_NAME}{The name of the first set (character string).}

\item{A}{The first set (vector).}

\item{B_NAME}{The name of the second set (character string).}

\item{B}{The second set (vector).}

\item{MAIN}{The main title (character string).}

\item{SUB}{The subtitle under the main title (character string).

The venn diagram is plotted in grid.draw().}
}
\value{
Plots the diagram without returning a value.
}
\description{
Venn diagram beween two sets A and B
}
\details{
This is a simple function wrapping \code{venn.diagram()} function from
the VennDiagram package, with a standard arbitrary look and feel.
}
\examples{
\dontrun{
VennAB( "Name of group A", groupA
      , "Name of group B", groupB
      , "Main title"
      , "subtitle")
}

}

